Address review feedback: simplify with slices, merge tests into handler_test.go
- expectedStatus is now []int, defaulted to [http.StatusOK] at
  construction time instead of checked lazily; isExpectedStatus removed
  in favor of slices.Contains at the call site.
- matchFallback rewritten using slices.IndexFunc instead of a manual
  loop.
- Fallback tests moved into the existing handler_test.go instead of a
  separate file.
